# Optimization: select_related when extracting vfolder from path