This makes a small difference, but nevertheless reduces load on db - which is a good thing.

Timings for /af/firefox/translate/user1/browser/#filter=incomplete:

OLD: 10 queries run, total 0.015 seconds
NEW: 8 queries run, total 0.008 seconds

also small reduction in the db time of get_units if vfolder in effect:

timings for /xhr/units/?path=%2Faf%2Ffirefox%2Fuser1%2Fbrowser%2F&initial=true&filter=incomplete:

OLD: 9 queries run, total 0.021 seconds
NEW: 7 queries run, total 0.018 seconds