@ta2-1 2016-01-21
lgtm after addressing nits.